Moving file rest api code back to storage app. Adding v6 scale file doc
