+++ scale/storage/serializers.py
@@ -4,8 +4,9 @@ from __future__ import unicode_literals
 import rest_framework.serializers as serializers
 from rest_framework.fields import CharField
 
+from batch.serializers import BatchBaseSerializerV6
+from recipe.serializers import RecipeTypeBaseSerializer
 from util.rest import ModelIdSerializer
-
 
 class DataTypeField(CharField):
     """Field for displaying the list of data type tags for a Scale file"""
@@ -155,3 +156,35 @@ class ScaleFileSerializerV6(ScaleFileBaseSerializerV6):
     center_point = WktField()
     meta_data = serializers.JSONField(default=dict)
     countries = serializers.StringRelatedField(many=True, read_only=True)
+
+class FileBaseSerializer(ScaleFileSerializerV6):
+    """Converts ScaleFile model fields to REST output"""
+    is_superseded = serializers.BooleanField()
+    superseded = serializers.DateTimeField()
+
+    source_started = serializers.DateTimeField()
+    source_ended = serializers.DateTimeField()
+
+    job_type = ModelIdSerializer()
+    job = ModelIdSerializer()
+    job_exe = ModelIdSerializer()
+    job_output = serializers.CharField()
+
+    recipe_type = ModelIdSerializer()
+    recipe = ModelIdSerializer()
+    recipe_job = serializers.CharField()
+    batch = ModelIdSerializer()
+
+
+class FileSerializer(FileBaseSerializer):
+    """Converts ScaleFile model fields to REST output"""
+    from job.serializers import JobTypeBaseSerializer
+
+    job_type = JobTypeBaseSerializer()
+    batch = BatchBaseSerializerV6()
+    recipe_type = RecipeTypeBaseSerializer()
+
+
+class FileDetailsSerializer(FileSerializer):
+    """Converts file model fields to REST output"""
+    pass
